#7に対応 ・階層クラスタリング→ＲａｎｄｏｍＦｏｒｅｓt ・共通関数を別ファイル化
